requirements.txt
dlib-bin==19.24.2
gdown==4.6.4
requests==2.31.0
Pillow==10.0.1
numpy==1.26.0